+++ src/Api/Installer/InstallerDescriptor.php
@@ -12,6 +12,7 @@
 namespace Puli\AssetPlugin\Api\Installer;
 
 use Puli\Manager\Assert\Assert;
+use Webmozart\Expression\Expression;
 
 /**
  * Describes a resource installer.
@@ -21,6 +22,21 @@ use Puli\Manager\Assert\Assert;
  */
 class InstallerDescriptor
 {
+    /**
+     * The name field in {@link Expression} instances.
+     */
+    const NAME = 'name';
+
+    /**
+     * The class name field in {@link Expression} instances.
+     */
+    const CLASS_NAME = 'className';
+
+    /**
+     * The description field in {@link Expression} instances.
+     */
+    const DESCRIPTION = 'description';
+
     /**
      * @var string
      */
@@ -240,5 +256,24 @@ class InstallerDescriptor
     public function hasParameterValue($parameterName)
     {
         return $this->hasParameter($parameterName) && !$this->getParameter($parameterName)->isRequired();
+    }
+
+    /**
+     * Returns whether the installer matches the given expression.
+     *
+     * @param Expression $expr The search criteria. You can use the fields
+     *                         {@link NAME}, {@link CLASS_NAME} and
+     *                         {@link DESCRIPTION} in the expression.
+     *
+     * @return bool Returns `true` if the installer matches the expression and
+     *              `false` otherwise.
+     */
+    public function match(Expression $expr)
+    {
+        return $expr->evaluate(array(
+            self::NAME => $this->name,
+            self::CLASS_NAME => $this->className,
+            self::DESCRIPTION => $this->description,
+        ));
     }
 }